Onboarding: assign launchpad-personalization at site creation, keep the chooser in all arms
The variation now resolves in the create-site step, so ai_launchpad sites are
created with wpcom_ai_launchpad_enabled (accepted by /sites/new server-side)
and every post-checkout path converges on the enabled launchpad. The Big Sky
chooser shows in all arms; treatments only replace the default My Home landing,
leaving the plugin, playground/blueprint, and Woo handoffs untouched. The
blank-site chooser exit goes straight to the variation destination.
