+++ client/landing/stepper/declarative-flow/flows/onboarding/test/onboarding.ts
@@ -3,8 +3,7 @@
  */
 import { renderHook } from '@testing-library/react';
 import { clearSessionStorageQuery } from 'calypso/components/domains/wpcom-domain-search/use-query-handler';
-import { loadExperimentAssignment } from 'calypso/lib/explat';
-import onboarding, { resolveLaunchpadPersonalizationVariation } from '../onboarding';
+import onboarding from '../onboarding';
 
 jest.mock( 'calypso/components/domains/wpcom-domain-search/use-query-handler', () => ( {
 	clearSessionStorageQuery: jest.fn(),
@@ -91,29 +90,5 @@ describe( 'onboarding flow side effects', () => {
 		renderSideEffect( 'domains' );
 
 		expect( clearSessionStorageQuery ).not.toHaveBeenCalled();
-	} );
-} );
-
-describe( 'resolveLaunchpadPersonalizationVariation', () => {
-	beforeEach( () => jest.clearAllMocks() );
-
-	it( 'forces ai_launchpad when the diy-launchpad override is present', async () => {
-		await expect( resolveLaunchpadPersonalizationVariation( '1' ) ).resolves.toBe( 'ai_launchpad' );
-		expect( loadExperimentAssignment ).not.toHaveBeenCalled();
-	} );
-
-	it( 'reads the variation from ExPlat when no override is present', async () => {
-		( loadExperimentAssignment as jest.Mock ).mockResolvedValue( {
-			variationName: 'no_guidance',
-		} );
-		await expect( resolveLaunchpadPersonalizationVariation( null ) ).resolves.toBe( 'no_guidance' );
-		expect( loadExperimentAssignment ).toHaveBeenCalledWith(
-			'wpcom_launchpad_personalization_202607_v1'
-		);
-	} );
-
-	it( 'falls back to control on an unrecognized assignment', async () => {
-		( loadExperimentAssignment as jest.Mock ).mockResolvedValue( { variationName: null } );
-		await expect( resolveLaunchpadPersonalizationVariation( null ) ).resolves.toBe( 'control' );
 	} );
 } );
