+++ client/landing/stepper/declarative-flow/internals/steps-repository/create-site/index.tsx
@@ -27,6 +27,7 @@ import Loading from 'calypso/components/loading';
 import useAddEcommerceTrialMutation from 'calypso/data/ecommerce/use-add-ecommerce-trial-mutation';
 import { useQuery } from 'calypso/landing/stepper/hooks/use-query';
 import { ONBOARD_STORE } from 'calypso/landing/stepper/stores';
+import { resolveLaunchpadPersonalizationVariation } from 'calypso/lib/ai-launchpad';
 import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
 import wpcom from 'calypso/lib/wp';
 import {
@@ -261,6 +262,14 @@ const CreateSite: StepType = function CreateSite( { navigation, flow, data } ) {
 		const isPlaygroundPublish =
 			sessionStorage.getItem( SESSION_KEY_FROM_PLAYGROUND_PUBLISH ) === '1';
 
+		// Assignment point for the launchpad-personalization experiment. Resolving the
+		// variation before creation lets ai_launchpad sites start with the AI Launchpad
+		// enabled, so every post-checkout path (direct, chooser, Big Sky return)
+		// converges on it regardless of which URLs the user actually visits.
+		const launchpadPersonalizationVariation = isOnboardingFlow( flow )
+			? await resolveLaunchpadPersonalizationVariation( urlQueryParams.get( 'diy-launchpad' ) )
+			: 'control';
+
 		const site = await createSite(
 			flow,
 			theme,
@@ -281,7 +290,9 @@ const CreateSite: StepType = function CreateSite( { navigation, flow, data } ) {
 			gardenName,
 			gardenPartnerName,
 			urlQueryParams.get( 'spec_id' ),
-			isPlaygroundPublish ? 'playground-publish' : undefined
+			isPlaygroundPublish ? 'playground-publish' : undefined,
+			undefined, // provisionTarget
+			launchpadPersonalizationVariation === 'ai_launchpad'
 		);
 
 		if ( ! site ) {
